[Refactoring ]Notification Modal 폴더구조 변경
