+++ src/components/NotificationModal/NotificationModal.Styles.ts
@@ -2,5 +2,8 @@ import styled from "styled-components"
 
 export const NotificationContainer = styled.div`
   padding: 1rem 2rem;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
 `
 export default NotificationContainer
