+++ src/components/NotificationModal/components/NotificationList/NotificationItem/NotificationItem.tsx
@@ -1,7 +1,7 @@
-import MessageProfile from "../../../pages/DirectMessage/components/MessageGroupList/MessageProfile"
-import * as S from "./Notification.Styles"
-import { Notification } from "../../../types/index"
-import getNotificationOption from "../utils/getNotificationOption"
+import MessageProfile from "@/pages/DirectMessage/components/MessageGroupList/MessageProfile"
+import * as S from "./NotificationItem.Styles"
+import { Notification } from "@/types"
+import getNotificationOption from "../../../utils/getNotificationOption"
 import { useLocation, useNavigate } from "react-router-dom"
 import { format, register } from "timeago.js"
 import kolocale from "timeago.js/lib/lang/ko"
@@ -11,24 +11,30 @@ interface NotificationItemProps {
   onClose: () => void
   Notification: Notification
 }
+
 const NotificationItem = ({
   handleClick,
   onClose,
-  Notification: notificationItem,
+  Notification: notification,
 }: NotificationItemProps) => {
-  const { createdAt, author } = notificationItem
+  const { createdAt, author } = notification
   const { pathname } = useLocation()
   const navigate = useNavigate()
-  const { NOTIFI_LANGUAGE, NOTIFI_NAVIGATION } = getNotificationOption(
-    notificationItem,
+  const { NOTIFI_TEXT, NOTIFI_NAVIGATION } = getNotificationOption(
+    notification,
     pathname,
   )
+
   register("ko", kolocale)
   const convertedData = format(createdAt, "ko")
 
   const handleClickProfileImage = () => {
     onClose()
     navigate(`/profile/${author._id}`)
+  }
+
+  const handleClickNotificationItem = () => {
+    handleClick(NOTIFI_NAVIGATION)
   }
 
   return (
@@ -39,16 +45,12 @@ const NotificationItem = ({
           profileImg={author.image}
           onClick={handleClickProfileImage}
         />
-        <S.NotificationItemInfo
-          onClick={() => {
-            handleClick(NOTIFI_NAVIGATION)
-          }}
-        >
+        <S.NotificationItemInfo onClick={handleClickNotificationItem}>
           <S.NotificationItemUserName>
             {author.fullName}
           </S.NotificationItemUserName>
           <S.NotificationItemContent>
-            {author.fullName}님이 {NOTIFI_LANGUAGE}
+            {author.fullName}님이 {NOTIFI_TEXT}
           </S.NotificationItemContent>
           <S.NotificationItemDate>{convertedData}</S.NotificationItemDate>
         </S.NotificationItemInfo>
